pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-moduflow"
version = "0.1.0"
description = "Moduflow environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core[core]>=0.2.2",
    "openai>=1.14.0",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "nltk>=3.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m moduflow.server.app
server = "moduflow.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["moduflow", "moduflow.server"]
package-dir = { "moduflow" = ".", "moduflow.server" = "server" }